# docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_interfaces_api_handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_interfaces_api_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_interfaces_api_handlers.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_interfaces_api_handlers.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  internal_interfaces_api_handlers.TokenResponse:
    properties:
      access_token:
        type: string
    type: object
  internal_interfaces_api_handlers.ValidateResponse:
    properties:
      message:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a microservice for user authentication.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Authentication Service API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user with email and password
      parameters:
      - description: User login data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_api_handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user and invalidate token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with name, email and password
      parameters:
      - description: User registration data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_api_handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/validate:
    get:
      consumes:
      - application/json
      description: Validate user token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.ValidateResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Validate token
      tags:
      - auth
schemes:
- http
swagger: "2.0"